requirements.txt
claude-agent-sdk==0.1.17
google-ads==28.4.1
typer>=0.9.0
python-dotenv==1.0.1
pydantic==2.5.3
pytest==9.0.2
pytest-cov==7.0.0
jinja2==3.1.4
instructor
anthropic
PyYAML==6.0.1
google-api-python-client==2.128.0
pytest-asyncio
mypy==1.8.0
ruff==0.1.8
pre-commit==3.6.0
black==23.12.1
